+++ packages/forms/src/zod/zod.ts
@@ -1,0 +1,4 @@
+import './i18n';
+import './schema';
+
+export * as z from 'zod';