+++ packages/forms/tests/unit/jest.config.ts
@@ -4,10 +4,10 @@ import { getTransformIgnorePatterns } from '@ttoss/test-utils';
 const config = jestUnitConfig({
   coverageThreshold: {
     global: {
-      statements: 94.57,
-      branches: 90.61,
-      functions: 95.63,
-      lines: 94.5,
+      statements: 94.18,
+      branches: 89.73,
+      functions: 94.3,
+      lines: 94.11,
     },
   },
   setupFilesAfterEnv: ['./setupTests.tsx'],