+++ packages/react-auth-cognito/tests/unit/jest.config.ts
@@ -4,9 +4,9 @@ import { getTransformIgnorePatterns } from '@ttoss/test-utils';
 const config = jestUnitConfig({
   coverageThreshold: {
     global: {
-      statements: 94,
-      branches: 80,
-      lines: 94,
+      statements: 91.5,
+      branches: 73.6,
+      lines: 91.5,
       functions: 85.71,
     },
   },