+++ packages/react-auth-core/tests/unit/jest.config.ts
@@ -4,10 +4,10 @@ import { getTransformIgnorePatterns } from '@ttoss/test-utils';
 const config = jestUnitConfig({
   coverageThreshold: {
     global: {
-      branches: 48,
-      functions: 50,
-      lines: 50,
-      statements: 50,
+      branches: 59.7,
+      functions: 61.4,
+      lines: 62.8,
+      statements: 62.8,
     },
   },
   setupFilesAfterEnv: ['./setupTests.tsx'],