+++ packages/react-auth-strapi/tests/unit/jest.config.ts
@@ -1,0 +1,20 @@
+import { jestUnitConfig } from '@ttoss/config';
+import { getTransformIgnorePatterns } from '@ttoss/test-utils';
+
+const config = jestUnitConfig({
+  coverageThreshold: {
+    global: {
+      branches: 0,
+      functions: 30,
+      lines: 25,
+      statements: 25,
+    },
+  },
+  setupFilesAfterEnv: ['./setupTests.tsx'],
+  testEnvironment: 'jsdom',
+  transformIgnorePatterns: getTransformIgnorePatterns({
+    esmModules: ['rehype-raw', 'react-error-boundary'],
+  }),
+});
+
+export default config;