+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.41.0](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.40.2...@ttoss/forms@0.41.0) (2026-02-15)
+
+### Features
+
+- react wizard ([#870](https://github.com/ttoss/ttoss/issues/870)) ([d309e78](https://github.com/ttoss/ttoss/commit/d309e78f798064c4e4e9a57cef94872c40f94d15))
+
 ## [0.40.2](https://github.com/ttoss/ttoss/compare/@ttoss/forms@0.40.1...@ttoss/forms@0.40.2) (2026-02-03)
 
 **Note:** Version bump only for package @ttoss/forms