+++ packages/react-notifications/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.6.0](https://github.com/ttoss/ttoss/compare/@ttoss/react-notifications@2.5.15...@ttoss/react-notifications@2.6.0) (2026-02-15)
+
+### Features
+
+- react wizard ([#870](https://github.com/ttoss/ttoss/issues/870)) ([d309e78](https://github.com/ttoss/ttoss/commit/d309e78f798064c4e4e9a57cef94872c40f94d15))
+
 ## [2.5.15](https://github.com/ttoss/ttoss/compare/@ttoss/react-notifications@2.5.14...@ttoss/react-notifications@2.5.15) (2026-02-03)
 
 **Note:** Version bump only for package @ttoss/react-notifications